Add support for custom headers and query parameters
Currently there is no way to send to the server custom headers or query
parameters. This patch adds support for that. For example, if the
caller, for whatever the reason, needs to send the 'My: myvalue' header
and the 'your=yourvalue' query parameter to the request to add a virtual
machine, it can be done as follows:

  # Get the reference to the service that managse the collection
  # of virtual machines:
  vms_service = connection.system_service().vms_service()

  # Add the virtual machine, with a custom query and header
  vms_service.add(
    types.Vm(),
    headers={'my': 'value'},
    query={'your': 'value'},
  )

Change-Id: I27ca5f421215da1d920474445be08746248e4aed
Signed-off-by: Ondra Machacek <omachace@redhat.com>
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>