Bump release number
Recent changes in the way that the automation scripts work had the side
effect of changing the RPM release number from 0.1 to 0. This means that
the new RPMs built in the automation environment, with RPM release 0,
are, from the RPM point of view, older than those with RPM release 0.1.
To avoid that this patch changes the RPM release number contained in the
.spec templates to 2, so that when decremented it will be 1, hence newer
than the existing 0.1 packages.

Change-Id: If64322d9a87fe6f099507e17076d17c84d6c5ba7
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>