Read correctly attributes with enum values
A recent patch introduced a mechanism to ignore incorrect values of
enum types. That patch also introduced a regresion for attributes of
enum types that have a name different to their type name. For
example, the 'vm.migration.auto_converge' attribute is of type
'InheritableBoolean', not of type 'AutoConverge' as assumed by the
code generator. This patch fixes that issue.

Change-Id: I7e8df50f4dd59881238346f7b9ce06dd2a653546
Signed-off-by: Ondra Machacek <omachace@redhat.com>
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>