Improve error message for wrong content type
Currently when the user specifies an incorrect URL path the SDK sends
requests to the server anyhow, and it tries to parse the response body
as XML, without checking the response content type. This generates error
messages that don't explain the root cause of the problem. To improve
that this patch changes the SDK so that it checks the content type of
the responses and generates error messages that explain better the
failure. For example, if the user specifies an incorrect URL path like
'/ugly', the SDK will try to contact the server and will raise an
exception containing the following error message:

   The response content type 'text/html; charset=iso-8859-1' isn't
   the expected XML. Is the path '/ugly' included in the 'url'
   parameter correct? The typical one is '/ovirt-engine/api'

Change-Id: I548ae621a025969fe2dbf086f49ba83ea52fc28f
Bug-Url: https://bugzilla.redhat.com/1440292
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Signed-off-by: Ondra Machacek <omachace@redhat.com>
(cherry picked from commit 5c1b391547a2c9d962818e0b31ba9bbb57b5e593)