Add generic writer
Currently the SDK doesn't have any mechanism to generate an XML document
from an object whose type isn't known. This patch adds a new
'Writer.write' method that decides what is the specific writer to use
based on the type of the object, so that the caller doesn't need to
explicitly select it.

Change-Id: I42d697b840a7b4ee2a197e881678e389f7cab919
Bug-Url: https://bugzilla.redhat.com/1444114
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>