Automatically replace bad tokens
Currently when the token used by the SDK is no longer valid, because it
is expired or for any other reason, the following requests just fail.
The only way to recover is to close the connection and open a new one.
That isn't convenient for long running applications. To improve that
this patch modifies the SDK so that it automatically replaces bad
tokens, requesting new ones from the SSO service.

Change-Id: I59e0dfef5a2b38b630a9c95ed19fea005acdf826
Bug-Url: https://bugzilla.redhat.com/1434830
Signed-off-by: Ondra Machacek <omachace@redhat.com>