Exclude unknown cURL debug types
Currently the cURL debug function that we use tries to convert all kinds
of debug data to strings, in order to send them to the log. This does
not work for binary data, for example for the SSL debug data. To avoid
that this patch changes the debug function so that it ignores the debug
data whose type doesn't know how to handle.

Change-Id: I26e811ebe93034e68ccc6cbb6f74f2314c6c6cfc
Bug-Url: https://bugzilla.redhat.com/1451042
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>