Add Error.fault
Currently when a fault is returned by the server the information that it
contains is used to generate an exception message, and then this
discarded. This patch changes the SDK so that the fault object is copied
to the 'Error.fault' attribute instead of discarded.

Change-Id: I2f8aac14214737824b8b51fac56d9a8657b299dd
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1450986
Signed-off-by: Ondra Machacek <omachace@redhat.com>
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 5c74d78c0b2351a796fd9bb19f5e7fe4386aa9d2)