Ignore unrelated responses from server
When using for example kerberos authentication the server can return
to the client more then one response. The client is interested only
in last response which contains the information about the
authentication, if it was successfull, so this patch add a logic to
ignore the headers which doesn't belong to last response.

Change-Id: I8025b471a9c05618f273c726a02a60017434ee58
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1459254
Signed-off-by: Ondra Machacek <omachace@redhat.com>
(cherry picked from commit 5b7458c618083ed4eb41b4a67fd0614724cb3fc1)