Fix parsing of multi-byte UTF-8 characters
Currently the Reader.read method assumes that strings need to be encoded
using UTF-8 before parsing them. This is only true in Python 3, where
strings are sequences of 16 bits characters. But it isn't true in Python
2, and trying to do so fails if the string is already encoded using
UTF-8. To avoid that this patch changes the Reader.read method so that
it only does this encoding when running with Python 3.

Change-Id: I6ab2764da91bfcd1bd50ebec228b2225eac2c446
Bug-Url: https://bugzilla.redhat.com/1461060
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>