No exceptions when 'raise_exception=False'
The contract of the 'Connection.test' method is that it doesn't raise
exceptions if the 'raise_exception' parameter is false. But that was
recently changed. This patch fixes that.

This patch also makes sure that the 'Connection.close' method does't
send any request if there is no token, as that is needed to make a
simple script like this work correctly when the connection does't work:

  # Create the connection, to a server that doesn't have the
  # oVirt engine:
  connection = sdk.Connection(
    ...
  )

  # Test the connection, which will fail, and return False:
  connection.test()

  # Close the connection, which should not try to release the token,
  # as there is no token to release:
  connection.close()

Change-Id: I64110106b7bd3ac60b2bc3692e41ac544d75ed67
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>