Get the disk to upload from the command line
Accept a path to the image, and check the image format. Based on the
format, configure the disk as raw or cow.

Once the user configure the connection details for the setup, and the
storage domain, the example code can be used for uploading disks to the
same setup without modifying anything.

Change-Id: Ib4a62aa6eb4164d1f67e95a85fd912412d803555
Signed-off-by: Nir Soffer <nsoffer@redhat.com>