Add support for method extensions
Currently the SDK doesn't support the methods that are defined in the
specification of the API as extensions of other methods. For example,
there is an `AddFromScratch` method that is an extension of the `Add`
method of the service that manages virtual machines. But the SDK doesn't
have a corresponding `add_from_scratch` method. This patch modifies the
code generator to add support for those method extensions.

Change-Id: I5d480d5f8a0878abb330fffe982a9576895bb9f9
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>