upload_disk: Add --disk-sparse argument
Previously we used:
- sparse for COW disk
- preallocated for RAW disk

But engine defaults are:
- block storage: RAW preallocated
- file storage: RAW sparse

So it was not possible to upload sparse raw files to file based storage.

It is not possible yet to use the same policy as engine, since we don't
have a way tell the type of the storage domain, and engine does not have
API to list available combinations or recommended combination for a
storage domain.

Until we have the required API, make it possible to create all
possible combinations when uploading disks. It would nice to block the
only unsupported combination (RAW sparse on block storage) but we don't
have any way to do this.

Change-Id: Ia491a53c7e924e96b482ead912f4921730b6ff93
Signed-off-by: Nir Soffer <nsoffer@redhat.com>