update CHANGES.adoc for release of 4.4.0
Note: descriptions of changes relating to automation in el8
have been ommitted, since they do not yet culminate to the
ability to run the automation in el8 (and thus would probably
be confusing). They will be added to CHANGES.adoc of the next
version. Also, changes to example files are not noted in
this file in order to keep it clean and readable

Change-Id: I1142415f6d68249e0d7773fac6b73ebf0d6671a3
Signed-off-by: Ori_Liel <oliel@redhat.com>