download_disk: Support download to qcow2 format
ovirt-imageio support now downloading to qcow2 images, since this is the
only format that can handle incremental backups.

Add argument to configure the download format so it can be used when
downloading complete disks. The default was changed to "qcow2" since
this format better for downloading disks, and upload_disk.py supports
now uploading qcow2 disks.

Change-Id: Ie81569a7d4f88a734b4ccf9391af0ba78f278e43
Signed-off-by: Nir Soffer <nsoffer@redhat.com>