{upload,download}_disk: Use proxy_url fallback
In ovirt-imageio 2.0.7 client.upload() and client.download() have
proxy_url argument. If set, the transfer will fallback to proxy_url if
transfer_url is not accessible.

Previously the transfer would fail, and need to run again with
--use-proxy. Now the transfer will just work.

I tested this by uploading and downloading from a machine without DNS
entries for the active hosts.

Change-Id: I074b84fc01b3c808a60c061d3a061766e59dac71
Signed-off-by: Nir Soffer <nsoffer@redhat.com>