backup_vm: Support backup via proxy
Since ovirt-imageio 2.0, we can backup also via proxy_url. We did not
use it since using the proxy is less efficient, but this will fail the
backup if transfer_url is not accessible.

Use the new proxy_url download parameter to make the backup work from
any host, even if transfer_url is not accessible. The fallback will work
with ovirt-imageio 2.0.7.

I tested this by backing up from a machine without DNS entries for the
active hosts.

Change-Id: Ie30d2898b7fc6b38a715cc36e3350c9cbcb3c32a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>