backup_vm: Validate VM UUID before starting backup
Previously if we used wrong vm_id, we failed to find any disks, and
called backup with empty disk list. This fails with a very confusing
error:

    ovirtsdk4.Error: Fault reason is "Input validation failed.". Fault
    detail is "Parameter 'backup.disks' is mandatory but was not
    provided.".  HTTP response code is 400.

This probably shows another issues in engine - the error should be
"backup.disk is empty". But the real issue is not the empty disk list
but trying to get disks from non-existing VM.

Now we check that the VM exists before starting the backup, so a wrong
vm_id will fail with a clear error:

    RuntimeError: VM 2d554b69-3c03-4f06-84b8-f4a7dd3756d3 does not exist

Change-Id: Ib4ef9c8e3f6cf8d4554b0acd8f3e89ca864f75fa
Signed-off-by: Nir Soffer <nsoffer@redhat.com>