{upload,download}_disk: Use local host if possible
If we are running on a hypervisor in the right data center and the
hypervisor is up, use the current host for the transfer. In this case we
can optimize the connection using unix socket, which speeds up the
transfer and reduce network bandwidth.

Finding the host is not trivial, and is needed in 4 examples, so I added
a helper module providing a find_host() function. We need to move more
shared code to this module to make the examples more clear and avoid
repeating boilerplate code.

Finding a host is not useful for backup, since we must use the host
where the vm is running. If you want to optimize backup you need to run
backup_vm.py on the host running the vm.

Change-Id: I688f5c998efaba3a9ab07b93928c07d7e95605d8
Signed-off-by: Nir Soffer <nsoffer@redhat.com>