upload_disk: Use imagetransfer helpers
When we started a transfer we did not wait correctly for the transfer,
and did not handle errors correctly, leaving leftover disks and paused
transfers on errors.

When finalizing a transfer we did not wait until the disk is finalized.
This always succeeds very quickly before the disk is unlocked, even when
upload verification failed and the uploaded disk was removed.

Most issues are fixed now by using imagetranster.create_transfer() and
imagetransfer.finalize_transfer().

Change-Id: I318e76750fd8656cc6f248c835e4d6ea4a9352c2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>