cancel_transfer: Add script to cancel a paused transfer
When the system paused a transfer, and the program that started the
transfer failed without canceling the transfer, the only way to cancel
the transfer and unlock the disk is to cancel it via the API.

In the past this was possible using the UI, but this was blocked since
the system does not support aborting image transfer. You can use this
script to cancel and a transfer when you know that the program that
started the transfer has disconnected from imageio server.

Change-Id: I6f8dc5a826f0299dce23b8261050a2c51938d340
Signed-off-by: Nir Soffer <nsoffer@redhat.com>