helpers: common: Make configure_logging more useful
Current code requires the caller to set the log level based on
the args.debug flag. It would be more useful if we can pass the parsed
arguments and let the common helper do the right thing.

Change the default logging level to INFO, same configuration used by
most examples. If we use a log file, we always want INFO log level.
WARNING level is good idea when we log to stderr.

Change-Id: If16a2c947e6d38f191dc178dfb027f2a0eb0c100
Signed-off-by: Nir Soffer <nsoffer@redhat.com>