readme: Document how to skip the tests
Building on Fedora 32 with recent flake8 is broken now because recent
flake8 is not happy with the current source. Document how to skip the
tests to make it possible to build and install on Fedora 32.

It would be better if the tests were separated from the build scripts
but current both check-patch.sh and build-artifacts use the automation
build.py.

Change-Id: I8ff2fdb088fb366433536e6ad887869344d62eaf
Signed-off-by: Nir Soffer <nsoffer@redhat.com>