helpers: units: Add units helper module
The module provides the common units to allow nicer code like:

    4 * MiB

Instead of:

    4 * 1024**2

It also provides a function for converting human size to bytes value
(e.g. "2m" -> 2 * MiB). This can be used as the type= argument when
parsing arguments.

Units letters are case insensitive, both "2m" and "2M" will be converted
to 2 * MiB.

Change-Id: I2e1da02acc7b280a2280434e4f333ea8c35d4b60
Signed-off-by: Nir Soffer <nsoffer@redhat.com>