checksum_disk: Show how to compute disk checksum
Add example using the new checksum API introduced in 2.0.10-1.

The example is minimal, showing only the basics. We can enhance it later
with more options like showing the available algorithms and selecting
the algorithm.

Since ImageioClient does not expose this API yet, we use standard
http.client.

Here is an example run:

$ ./checksum_disk.py -c engine3 58daea80-1229-4c6b-b33c-1a4e568c8ad7
{
    "checksum": "05953b09bb83ba258e027d202fc6dd3170c75d2f",
    "algorithm": "sha1"
}

Change-Id: I00af527d853f647d4823d2686c71839713e3188e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>