imagetransfer: Support download of disk snapshots
Add support for several use cases which were not possible when using the
imagetransfer helper.

1. Downloading disk data at the time of a snapshot.

User locate the wanted disk snapshot, and create a transfer with the
disk snapshot instead of the disk:

    transfer = imagetransfer.create_transfer(
        connection=connection,
        disk_snapshot=disk_snapshot,
        direction=types.ImageTransferDirection.DOWNLOAD,
    )

The transfer include the disk contents at the time of the snapshot,
including data from previous snapshots and the from a template if this
disk is based on a template.

2. Downloading a single disk snapshot.

The typical use case is snapshot based incremental backup. The user
creates a snapshot, and create a transfer with the disk snapshot,
specifying shallow=True:

    transfer = imagetransfer.create_transfer(
        connection=connection,
        disk_snapshot=disk_snapshot,
        direction=types.ImageTransferDirection.DOWNLOAD,
        shallow=True,
    )

The transfer includes data only from the specified disk snapshot.
Unallocated areas in the snapshots are reported as holes when using the
extents API in imageio.

3. Downloading the disk data from the current snapshot.

We don't expose a disk snapshot for the active disk snapshot. To
download it, the user can specify the disk object, and shallow=True.

    transfer = imagetransfer.create_transfer(
        connection=connection,
        disk=disk,
        direction=types.ImageTransferDirection.DOWNLOAD,
        shallow=True,
    )

The transfer includes only data from the active disk snapshot of the
disk.

Change-Id: Ie99270a89ad498cfd7c8b3c3109e6e4cbc526015
Bug-Url: https://bugzilla.redhat.com/1847090
Signed-off-by: Nir Soffer <nsoffer@redhat.com>