change_vm_cd: use common helpers
Use common helper in example for changing VM CD. Besides using common
config file, it also passes VM name and disk ID as parameters to the
script, so the user don't have to read whole source code and modify it.

Also changes the default value for `current` parameter to `True`, so
that CD is by default changed immediately, not after next VM reboot.

Example usage for immediate change (but ephemeral):

    ./change_vm_cd.py -c engine1 --disk-id 02b5dc5b-2bdc-4c28-8a36-790e8c6cb09c test-vm

or, for changing CD after next reboot (but permanent change):

    ./change_vm_cd.py -c engine1 --disk-id 02b5dc5b-2bdc-4c28-8a36-790e8c6cb09c --permanent test-vm

Change-Id: Idc27668242ce328d6b008dee333614e7401708a9
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>