backup_vm: Improve logging when finalizing backup
Add "Finalizing backup" and "Waiting until backup is finalized" logs to
make finalize flow more clear. Previously we did not have any logs for
this step during full and incremental backup.

Here is an example showing the new logs:

$ ./backup_vm.py -c engine-dev full --backup-dir /var/tmp/backups/raw 4dc3bb16-f8d1-4f59-9388-a93f68da7cf0
[   0.0 ] Starting full backup for VM 4dc3bb16-f8d1-4f59-9388-a93f68da7cf0
[   0.8 ] Waiting until backup f01630c5-943c-4845-b404-29e5ff0b509b is ready
[   1.8 ] Creating image transfer for disk 126eea31-c5a2-4c01-a18d-9822b0c05c2a
[   3.1 ] Image transfer 2c150870-915b-4cc2-a4a1-ca32287099de is ready
[  45.55% ] 2.73 GiB, 2.84 seconds, 983.96 MiB/s
[   5.9 ] Finalizing image transfer
[   9.0 ] Finalizing backup
[  39.1 ] Waiting until backup is finalized
...

We can see that finalizing backup took about 30 seconds,
and then we are stuck waiting until the backup is finalized.

Change-Id: Ic55279feec109070e553765009d1332187c1bdc7
Signed-off-by: Nir Soffer <nsoffer@redhat.com>