python-sdk: added timeout_policy attribute
Timeout ploicy defines the action to do,
after the transfer's inactivity timeout
is passed.

The new --timeout-policy argument has three options:
legacy, pause and cancel.

The default option for the updated example scripts
is 'cancel'.

After inactivity timeout, if the option is 'cancel',
then, the transfer will be canceled. And if the option
'pause', then, the transfer will be pauesd. The 'legay'
option cancels the transfer if the direction is download,
and, pauses it if it is upload.

Bug-Url: https://bugzilla.redhat.com/1892676
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>
Change-Id: I3945e17dc879339cc1ab295519dadd33e896beb6