Preserve /dev/null when running under root user
It seems that Python SDK code running under root user is overwriting
/dev/null file, when this file is passed as a parameter COOKIEJAR
to pycurl.

According to the documentation we could achieve the desired behavior
(don't load any cookies at the beginning and don't store any cookies
to a file during connection) by omitting the COOKIEJAR initialization
and initializing COOKIEFILE to "".

Change-Id: I9d61dda53c02f39458143a895544c58623949bd8
Bug-Url: https://bugzilla.redhat.com/1956750
Signed-off-by: Martin Perina <mperina@redhat.com>