download_disk: Fix error handling
When creating a transfer, the client must finalize or cancel the
transfer. Previously, an exception raised after the transfer was started
would leave the transfer unclosed. This is a very bad since it leaves
prepared images on the host, and running nbd server, and it is also a
horrible example.

Move the try block under the line creating the transfer, so any error
after that will cause the transfer to finalize.

Change-Id: I7c04f4f46edd969909f8f5e6a3cc2ee1cdcad765
Signed-off-by: Nir Soffer <nsoffer@redhat.com>