upload_disk: Fix error handling
Similar to download_disk, we must finalize the transfer after successful
download. But unlike download, when we upload data to a new disk, we
should cancel the transfer on errors. This will delete the new disk with
the partial data, without wasting time on verifying the uploaded disk.

Fix the example to use proper error handling so it can server as example
code.

Change-Id: I6e0c8ae64ac00554b39f154991ee0ad65df03384
Signed-off-by: Nir Soffer <nsoffer@redhat.com>