upload_disk: Add --disk-id option
Add --disk-id option so a user can specify the new disk id for the
upload. This is useful mainly for testing and can also be used for
reconstructing VM snapshots from backups with the same disk UUIDs.

With this, we can verify that engine locks are released properly when
upload is finalized, and we can start download right after that. In the
past this used to fail randomly.

Here is an example run using the shiny --disk-id:

$ ./upload_disk.py -c engine-dev \
        --sd-name iscsi-01 \
        --disk-sparse \
        --disk-format qcow2 \
        --disk-id b9be4fd5-05ae-4264-9f91-6830cb638886 \
        fedora-32.qcow2; \
    ./download_disk.py -c engine-dev \
        b9be4fd5-05ae-4264-9f91-6830cb638886 \
        /tmp/download.qcow2
[   0.0 ] Checking image...
[   0.0 ] Image format: qcow2
[   0.0 ] Disk format: cow
[   0.0 ] Disk content type: data
[   0.0 ] Disk provisioned size: 6442450944
[   0.0 ] Disk initial size: 1661075456
[   0.0 ] Disk name: fedora-32.qcow2
[   0.0 ] Disk backup: False
[   0.0 ] Connecting...
[   0.0 ] Creating disk...
[  17.3 ] Disk ID: b9be4fd5-05ae-4264-9f91-6830cb638886
[  17.3 ] Creating image transfer...
[  19.1 ] Transfer ID: bab30c5f-37a9-47fa-acb6-212b789cf1cc
[  19.1 ] Transfer host name: host3
[  19.1 ] Uploading image...
[ 100.00% ] 6.00 GiB, 10.12 seconds, 607.25 MiB/s
[  29.2 ] Finalizing image transfer...
[  32.2 ] Upload completed successfully
[   0.0 ] Connecting...
[   0.1 ] Creating image transfer...
[   1.7 ] Transfer ID: 0bd779ea-afc3-4417-9375-c38a01dabab7
[   1.7 ] Transfer host name: host4
[   1.7 ] Downloading disk...
[ 100.00% ] 6.00 GiB, 3.32 seconds, 1.81 GiB/s
[   5.0 ] Finalizing image transfer...
[   7.0 ] Download completed successfully

Running this in a loop is a good test for virt-v2v flow, when right
after disk is uploaded, we try to attach the disk to a vm.

Change-Id: I4332ae518837e69f58d97ecf8b3cc13a4badc8d2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>