backup: fix misleading RuntimeError message
In case of failed backup, the "RuntimeError" raised by
"get_backup()" included the last backup event that was
supposed to shed some light on the failure reason.
In practice, most of the times it was confusing, since it
pointed to a non-related event as a reason for failure, like:
"Backup ... for VM ... started"
or
"The disk ... was successfully added".
Updating the message phrasing, so it will be clear that the event is
simply the last reported event, i.e., "best effort" style to help user.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Change-Id: Id085c5d8283460338db1224c782b71c5b781bd9c