Switch from Travis to GitHub Actions
After Travis moved domains from .org to .com they have also changed
billing plan so we have decided to move to GitHub Actions.
This patch adds build workflow which will build the sdk and
documentation. Created sdk will be pushed to new repository [1] and the
documentation will be uploaded to the GitHub Pages.
[1] https://www.github.com/ovirt/python-ovirt-engine-sdk4

Signed-off-by: Martin Necas <mnecas@redhat.com>
Change-Id: Ic4f6ae87ae9b81eab95b19efdbffa878d77f4073
Signed-off-by: Martin Necas <mnecas@redhat.com>