doc: use reStructuredText for python arguments
Previously the docs generated via pdoc used asciidoc format.
But the asciidoc support in pdoc was added via a custom patch [1].
As we don't want to keep that up-to-date with pdoc versions, we move
away from asciidoc to reStructuredText which is the default for pdoc.

The main change is in the arguments definition, which goes from:
`argumentname`: description
to
:param argumentname: description

[1]: https://github.com/machacekondra/pdoc

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>