Dockerfile: use buildcontainer for devcontainer
Use our centos stream 9 buildcontainer as a base for the devcontainer
instead of an Ubuntu image.
Also use Java21 like we use in the spec.

Next to that, run as 'build' user within the container.
This the avoid running as root within the container and having
permissions issues with the local machine.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>